+++ repository/favorite.go
@@ -16,12 +16,12 @@ func NewFavoriteRepository(db *gorm.DB) *FavoriteRepository {
 }
 
 //Create Создание
-func (repo *FavoriteRepository) Create(userID int, apartID int) error {
+func (repo *FavoriteRepository) Create(userID int64, apartID int) error {
 	return repo.Exec("INSERT INTO favorite_apartments (user_id, apartment_id) VALUES(?, ?);", userID, apartID).Error
 }
 
 //Remove удаление записи
-func (repo *FavoriteRepository) Remove(userID int, apartID int) (bool, error) {
+func (repo *FavoriteRepository) Remove(userID int64, apartID int) (bool, error) {
 	var count int
 	err := repo.Raw(
 		"WITH deleted AS (DELETE FROM favorite_apartments  WHERE user_id = ? AND apartment_id = ? RETURNING *) SELECT count(*) FROM deleted;",
@@ -38,7 +38,7 @@ func (repo *FavoriteRepository) Remove(userID int, apartID int) (bool, error) {
 }
 
 //FindFirst проверить наличие такой же записи
-func (repo *FavoriteRepository) FindFirst(userID int, apartID int) (bool, error) {
+func (repo *FavoriteRepository) FindFirst(userID int64, apartID int) (bool, error) {
 	var count int
 	err := repo.Raw(
 		"SELECT count(*) FROM favorite_apartments f WHERE f.user_id = ? AND f.apartment_id = ? LIMIT 1;", userID, apartID).Count(&count).Error
@@ -52,7 +52,7 @@ func (repo *FavoriteRepository) FindFirst(userID int, apartID int) (bool, error)
 }
 
 //All Найти избранную недвижимоть пользователя
-func (repo *FavoriteRepository) All(userID int, aparts *[]models.Apartment) error {
+func (repo *FavoriteRepository) All(userID int64, aparts *[]models.Apartment) error {
 	rows, err := repo.Raw(
 		`SELECT * FROM apartment a
 		LEFT JOIN favorite_apartments f ON a.id = f.apartment_id
